requirements.txt
Django==1.7
Markdown==2.5.1
MySQL-Connector-Python==1.2.2
South==1.0
WebOb==1.4
WebTest==2.0.16
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.18
celery==3.1.16
certifi==14.05.14
colorama==0.3.2
coverage==3.7.1
django-autofixture==0.9.1
django-celery==3.1.16
django-common-helpers==0.6.2
django-concurrency==0.8.1
django-cron==0.3.4
django-filter==0.8
django-webtest==1.7.7
djangorestframework==2.4.3
docutils==0.11
kombu==3.0.23
mando==0.3.2
mccabe==0.2.1
pytz==2014.7
radon==1.1.1
six==1.8.0
virtualenv==1.11.4
waitress==0.8.9
